requirements.txt
discord.py
python-dotenv
httpx
requests
